+++ R/removeCluster.R
@@ -169,9 +169,18 @@ removeClusterST <- function(area,
   }
   
   dirs_to_remove <- file.path(clustertypePath, subdirs_to_remove, area)
+  
+  # remove dir "constraints" (st-storage)
+  if(opts$antaresVersion>=920)
+    dirs_to_remove <- file.path(clustertypePath, 
+                                append(subdirs_to_remove, "constraints"), 
+                                area)
+  
+  # if list.ini is empty, area/series area/constraints can be removed
   if (length(previous_params) > 0) {
     dirs_to_remove <- file.path(dirs_to_remove, cluster_name)
   }
+  
   lapply(dirs_to_remove, unlink, recursive = TRUE)
   
   # Maj simulation