+++ tests/testthat/test-RemoveClusterST.R
@@ -179,7 +179,7 @@ suppressWarnings(
 area_test_clust = "al" 
 createArea(name = area_test_clust)
 
-test_that("Remove cluster", {
+test_that("Remove cluster without constraints", {
   # at least need 1 st cluster
   createClusterST(area = area_test_clust, 
                   cluster_name = "cluster_init")
@@ -190,6 +190,65 @@ test_that("Remove cluster", {
   
   # removed from ini file ?
   
+  # read prop
+  path_st_ini <- file.path("input", 
+                           "st-storage", 
+                           "clusters", 
+                           area_test_clust,
+                           "list")
+  
+  read_ini <- antaresRead::readIni(path_st_ini)
+  target_prop <- read_ini[[paste(area_test_clust, 
+                                 "cluster_init",
+                                 sep = "_")]]
+  
+  expect_null(target_prop)
+  
+  # remove directory ? 
+  path_st_dir <- file.path("input", 
+                           "st-storage", 
+                           "clusters", 
+                           "series",
+                           area_test_clust,
+                           "al_cluster_init")
+  
+  expect_false(dir.exists(path_st_dir))
+})
+
+
+test_that("Remove cluster with constraints", {
+  # given
+  name_no_prefix <- "remove_constraints"
+  
+  constraints_properties <- list(
+    "withdrawal-1"=list(
+      variable = "withdrawal",
+      operator = "equal",
+      hours = c("[1,3,5]", 
+                "[120,121,122,123,124,125,126,127,128]")
+    ),
+    "netting-1"=list(
+      variable = "netting",
+      operator = "less",
+      hours = c("[1, 168]")
+    ))
+  
+  good_ts <- data.table::as.data.table(matrix(10, 8760))
+  constraints_ts <- list(
+    "withdrawal-1"=good_ts,
+    "netting-1"=good_ts)
+  
+  createClusterST(area = area_test_clust, 
+                  cluster_name = name_no_prefix, 
+                  constraints_properties = constraints_properties, 
+                  constraints_ts = constraints_ts)
+  
+  # when
+  # remove cluster
+  removeClusterST(area = area_test_clust, 
+                  cluster_name = name_no_prefix)
+  
+  # then
   # read prop
   path_st_ini <- file.path("input", 
                            "st-storage",